lesson about gates

+++ app/Http/Controllers/ProductController.php
@@ -14,6 +14,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        $this->authorize("view", "products");
         return ProductResource::collection(Product::paginate());
     }
 
@@ -22,6 +23,7 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("edit", "products");
         $product = Product::create($request->only("title", "description", "image", "price"));
         return response(new ProductResource($product), Response::HTTP_CREATED);
     }
@@ -31,6 +33,7 @@ class ProductController extends Controller
      */
     public function show(int $id)
     {
+        $this->authorize("view", "products");
         return new ProductResource(Product::find($id));
     }
 
@@ -39,6 +42,7 @@ class ProductController extends Controller
      */
     public function update(Request $request, int $id)
     {
+        $this->authorize("edit", "products");
         $product = Product::find($id);
         $product->update($request->only("title", "description", "image", "price"));
         return response(new ProductResource($product), Response::HTTP_ACCEPTED);
@@ -49,6 +53,7 @@ class ProductController extends Controller
      */
     public function destroy(int $id)
     {
+        $this->authorize("edit", "products");
         Product::destroy($id);
         return response(null, Response::HTTP_NO_CONTENT);
     }
