+++ app/Http/Controllers/UserController.php
@@ -15,12 +15,14 @@ class UserController extends Controller
 
     public function index()
     {
+        $this->authorize("view", "users");
         return UserResource::collection(User::with('role')->paginate(5));
     }
 
 
     public function store(UserCreateRequest $request)
     {
+        $this->authorize("edit", "users");
         $user = User::create(
             $request->only("first_name", "last_name", "email", "role_id")
             + ["password" => Hash::make("1234")]
@@ -32,12 +34,14 @@ class UserController extends Controller
 
     public function show(string $id)
     {
+        $this->authorize("view", "users");
         return new UserResource(User::with('role')->find($id));
     }
 
 
     public function update(UserUpdateRequest $request, string $id)
     {
+        $this->authorize("edit", "users");
         $user = User::find($id);
         $user->update($request->only("first_name", "last_name", "email", "role_id"));
         return response(new UserResource($user), Response::HTTP_ACCEPTED);
@@ -46,6 +50,7 @@ class UserController extends Controller
 
     public function destroy(string $id)
     {
+        $this->authorize("edit", "users");
         User::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
